feat: adiciona comando de execução do ffmpeg
